+++ packages/shared/widget-plugin-filtering/src/controllers/input/NumberInputController.ts
@@ -1,9 +1,9 @@
 import { debounce } from "@mendix/widget-plugin-platform/utils/debounce";
 import { action, autorun, computed, makeObservable, reaction, runInAction } from "mobx";
 import { createRef } from "react";
-import { InputStore } from "../stores/input/InputStore";
-import { FilterFunctionBinary, FilterFunctionGeneric, FilterFunctionNonValue } from "../typings/FilterFunctions";
-import { FilterV, Number_InputFilterInterface } from "../typings/InputFilterInterface";
+import { InputStore } from "../../stores/input/InputStore";
+import { FilterFunctionBinary, FilterFunctionGeneric, FilterFunctionNonValue } from "../../typings/FilterFunctions";
+import { FilterV, Number_InputFilterInterface } from "../../typings/InputFilterInterface";
 
 export type Params = {
     filter: Number_InputFilterInterface;
